+++ x/attribute/types/msgs_test.go
@@ -21,15 +21,16 @@ var (
 // test ValidateBasic for TestMsgAddAttribute
 func TestMsgAddAttribute(t *testing.T) {
 	tests := []struct {
-		account, owner     sdk.AccAddress
+		account            string
+		owner              sdk.AccAddress
 		name, proposalType string
 		proposalValue      string
 		expectPass         bool
 	}{
-		{nil, addrs[1], "test", "string", "nil owner", false},
-		{addrs[0], nil, "test", "string", "nil account", false},
-		{nil, nil, "test", "string", "nil owner and account", false},
-		{addrs[0], addrs[1], "test", "string", "valid attribute", true},
+		{"", addrs[1], "test", "string", "nil owner", false},
+		{addrs[0].String(), nil, "test", "string", "nil account", false},
+		{"", nil, "test", "string", "nil owner and account", false},
+		{addrs[0].String(), addrs[1], "test", "string", "valid attribute", true},
 	}
 
 	for i, tc := range tests {
@@ -54,18 +55,19 @@ func TestMsgAddAttribute(t *testing.T) {
 // test ValidateBasic for TestMsgUpdateAttribute
 func TestMsgUpdateAttribute(t *testing.T) {
 	tests := []struct {
-		account, owner sdk.AccAddress
-		name           string
-		originalValue  []byte
-		originalType   AttributeType
-		updateValue    []byte
-		updateType     AttributeType
-		expectPass     bool
-		expectedError  string
+		account       string
+		owner         sdk.AccAddress
+		name          string
+		originalValue []byte
+		originalType  AttributeType
+		updateValue   []byte
+		updateType    AttributeType
+		expectPass    bool
+		expectedError string
 	}{
-		{addrs[0], addrs[1], "example", []byte("original"), AttributeType_String, []byte("update"), AttributeType_Bytes, true, ""},
-		{nil, addrs[1], "example", []byte("original"), AttributeType_String, []byte("update"), AttributeType_Bytes, false, ""},
-		{addrs[0], nil, "example", []byte(""), AttributeType_String, []byte("update"), AttributeType_Bytes, false, ""},
+		{addrs[0].String(), addrs[1], "example", []byte("original"), AttributeType_String, []byte("update"), AttributeType_Bytes, true, ""},
+		{"", addrs[1], "example", []byte("original"), AttributeType_String, []byte("update"), AttributeType_Bytes, false, ""},
+		{addrs[0].String(), nil, "example", []byte(""), AttributeType_String, []byte("update"), AttributeType_Bytes, false, ""},
 	}
 
 	for _, tc := range tests {
@@ -82,15 +84,16 @@ func TestMsgUpdateAttribute(t *testing.T) {
 // test ValidateBasic for TestMsgDeleteDistinctAttribute
 func TestMsgDeleteDistinctAttribute(t *testing.T) {
 	tests := []struct {
-		account, owner sdk.AccAddress
-		name           string
-		value          []byte
-		attrType       AttributeType
-		expectPass     bool
+		account    string
+		owner      sdk.AccAddress
+		name       string
+		value      []byte
+		attrType   AttributeType
+		expectPass bool
 	}{
-		{addrs[0], addrs[1], "example", []byte("original"), AttributeType_String, true},
-		{nil, addrs[1], "example", []byte("original"), AttributeType_String, false},
-		{addrs[0], nil, "example", []byte(""), AttributeType_String, false},
+		{addrs[0].String(), addrs[1], "example", []byte("original"), AttributeType_String, true},
+		{"", addrs[1], "example", []byte("original"), AttributeType_String, false},
+		{addrs[0].String(), nil, "example", []byte(""), AttributeType_String, false},
 	}
 
 	for _, tc := range tests {